+++ hutool-core/src/main/java/cn/hutool/core/img/Img.java
@@ -277,16 +277,26 @@ public class Img {
 	 * @return this
 	 */
 	public Img pressImage(Image pressImg, int x, int y, float alpha) {
-		final BufferedImage destImg = this.srcImage;
-		final int width = destImg.getWidth();
-		final int height = destImg.getHeight();
-
-		int pressImgWidth = pressImg.getWidth(null);
-		int pressImgHeight = pressImg.getHeight(null);
-		x += (width - pressImgWidth) / 2;
-		y += (height - pressImgHeight) / 2;
+		final int pressImgWidth = pressImg.getWidth(null);
+		final int pressImgHeight = pressImg.getHeight(null);
 		
-		draw(destImg, pressImg, new Rectangle(x, y, pressImgWidth, pressImgHeight), alpha);
+		return pressImage(pressImg, new Rectangle(x, y, pressImgWidth, pressImgHeight), alpha);
+	}
+	
+	/**
+	 * 给图片添加图片水印<br>
+	 * 此方法并不关闭流
+	 * 
+	 * @param pressImg 水印图片，可以使用{@link ImageIO#read(File)}方法读取文件
+	 * @param rectangle 矩形对象，表示矩形区域的x，y，width，height，x,y从背景图片中心计算
+	 * @param alpha 透明度：alpha 必须是范围 [0.0, 1.0] 之内（包含边界值）的一个浮点数字
+	 * @return this
+	 * @since 4.1.14
+	 */
+	public Img pressImage(Image pressImg,Rectangle rectangle, float alpha) {
+		final BufferedImage destImg = this.srcImage;
+		
+		draw(destImg, pressImg, rectangle, alpha);
 		this.destImage = destImg;
 		return this;
 	}
@@ -390,12 +400,21 @@ public class Img {
 	 * 
 	 * @param backgroundImg 背景图片
 	 * @param img 要绘制的图片
-	 * @param rectangle 矩形对象，表示矩形区域的x，y，width，height
+	 * @param rectangle 矩形对象，表示矩形区域的x，y，width，height，x,y从背景图片中心计算
 	 * @return 绘制后的背景
 	 */
 	private static BufferedImage draw(BufferedImage backgroundImg, Image img, Rectangle rectangle, float alpha) {
 		final Graphics2D g = backgroundImg.createGraphics();
 		g.setComposite(AlphaComposite.getInstance(AlphaComposite.SRC_ATOP, alpha));
+		
+		//修正图片位置从背景的中心计算
+		final int width = backgroundImg.getWidth();
+		final int height = backgroundImg.getHeight();
+		rectangle.setLocation(//
+				rectangle.x + (int)(Math.abs(width - rectangle.width) / 2), //
+				rectangle.y + (int)(Math.abs(height - rectangle.height) / 2)//
+		);
+		
 		g.drawImage(img, rectangle.x, rectangle.y, rectangle.width, rectangle.height, null); // 绘制切割后的图
 		g.dispose();
 		return backgroundImg;