+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core】         StrUtil增加hide方法
 * 【core】         PatternPool增加URL_HTTP，原URL规则变更
 * 【extra】        统一FTP和SFTP接口规范
+* 【extra】        QrCodeUtil支持二维码中贴Logo图片
 
 ### Bug修复
 * 【core】         修复ImageUtil.convert转换png变色问题（issue#IMWUO@Gitee）